feat: styled table + compartmentalised table
